qlqw_ng: simplify error handling
This change does away with the complicated delayed message-passing from
the initial implementation. As errorhook() is safe to call from any
Thread instance, prefer instead to call it directly from the I/O thread
if anything bad happens. The main plugin class itself is now intended
to be a bulletproof wrapper which does not fail in any meaningful sense.